requirements.txt
beautifulsoup4==4.4.0
bleach==1.4.1
chardet==2.3.0
Django==1.8.4
django-bootstrap3==6.2.2
django-ckeditor==4.5.1
django-contrib-comments==1.6.1
django-tinymce==2.0.4
filebrowser-safe==0.4.0
future==0.15.0
grappelli-safe==0.4.1
html5lib==0.999999
Mezzanine==4.0.1
mysqlclient==1.3.6
oauthlib==1.0.3
Pillow==2.9.0
pytz==2015.4
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
tzlocal==1.2
